SPACommentsBackend/requirements.txt
django == 4.1.7
djangorestframework == 3.14.0
django-cors-headers == 3.14.0
drf-spectacular == 0.26.0
gunicorn == 20.1.0